minor refactor: added debug constexpr, static_cast to reduce warnings

+++ src/Core/HUD.cpp
@@ -7,7 +7,7 @@ HUD::HUD() {
 }
 
 void HUD::UpdateHUD() {
-    healthBar.width = playerCharacter->GetHealth() * 2;
+    healthBar.width = static_cast<float>(playerCharacter->GetHealth() * 2);
 }
 
 void HUD::DrawHUD() {
